requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaggle>=1.5.0
huggingface_hub>=0.17.0
pyyaml>=6.0
tqdm>=4.66.0
scipy>=1.11.0
torchmetrics>=1.0.0
accelerate>=0.23.0
python-dotenv>=1.0.0
